chk_pkg_dep
